concatS in separate module
